+++ confiture-rest-api/prisma/migrations/20251204095851_add_statement_publication_and_edition_dates/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "Audit" ADD COLUMN     "statementEditionDate" TIMESTAMP(3),
+ADD COLUMN     "statementPublicationDate" TIMESTAMP(3);