+++ confiture-rest-api/src/audits/audit.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import _, { omit, orderBy, partition, pick, setWith, sortBy, uniqBy } from "lodash";
+import _, { isEqual, omit, orderBy, partition, pick, setWith, sortBy, uniqBy } from "lodash";
 import { nanoid } from "nanoid";
 import sharp from "sharp";
 import {
@@ -486,7 +486,18 @@ export class AuditService {
       // update audit edition date only if a property other than below has been changed
       const ignoredChanges: (keyof typeof audit)[] = ["auditorName", "procedureName", "auditorEmail"];
       if (!changedProperties.every(changedProperty => ignoredChanges.includes(changedProperty))) {
-        return (await this.updateAuditEditDate(uniqueId)) ?? audit;
+        await this.updateAuditEditDate(uniqueId);
+      }
+
+      // Update statement date when `procedureName` or `pages` change and if there is a `statementPublicationDate`
+      if (
+        audit.statementPublicationDate &&
+        (
+          !isEqual(previousAudit.pages, audit.pages ||
+          changedProperties.includes("procedureName"))
+        )
+      ) {
+        return (await this.updateStatementDate(uniqueId)) ?? audit;
       }
 
       return audit;
@@ -569,6 +580,35 @@ export class AuditService {
       ...promises
     ]);
     await this.updateAuditEditDate(uniqueId);
+
+    // Update `statementPublicationDate` on:
+    // - a single criterium status update
+    // - multiple criteria updates
+    const criterium = await this.prisma.criterionResult.findUnique({
+      where: {
+        pageId_topic_criterium: {
+          criterium: body.data[0].criterium,
+          topic: body.data[0].topic,
+          pageId: body.data[0].pageId
+        }
+      },
+      select: {
+        status: true
+      }
+    });
+
+    const statusChanged = body.data.length === 1 && criterium.status !== body.data[0].status;
+
+    if (body.data.length > 1 || statusChanged) {
+      const audit = await this.prisma.audit.findUnique({
+        where: { editUniqueId: uniqueId },
+        select: { statementPublicationDate: true }
+      });
+
+      if (audit.statementPublicationDate) {
+        this.updateStatementDate(uniqueId);
+      }
+    }
   }
 
   async saveExampleImage(
@@ -882,7 +922,6 @@ export class AuditService {
         },
         data: {
           publicationDate: new Date()
-          // editionDate: null,
         },
         include: AUDIT_EDIT_INCLUDE
       });
@@ -910,6 +949,22 @@ export class AuditService {
         include: AUDIT_EDIT_INCLUDE
       });
     }
+  }
+
+  // Either update statement publication or edition date
+  private async updateStatementDate(uniqueId: string) {
+    const audit = await this.prisma.audit.findUnique({
+      where: { editUniqueId: uniqueId },
+      select: { statementPublicationDate: true }
+    });
+
+    return this.prisma.audit.update({
+      where: { editUniqueId: uniqueId },
+      data: audit.statementPublicationDate
+        ? { statementEditionDate: new Date() }
+        : { statementPublicationDate: new Date() },
+      include: AUDIT_EDIT_INCLUDE
+    });
   }
 
   async getAuditReportData(
@@ -1013,6 +1068,8 @@ export class AuditService {
       creationDate: audit.creationDate,
       publishDate: audit.publicationDate,
       updateDate: audit.editionDate,
+      statementPublicationDate: audit.statementPublicationDate,
+      statementEditionDate: audit.statementEditionDate,
 
       notCompliantContent: audit.notCompliantContent,
       derogatedContent: audit.derogatedContent,
@@ -1412,6 +1469,16 @@ export class AuditService {
         creationDate: new Date(),
         editionDate: undefined,
         publicationDate: originalAudit.publicationDate ? new Date() : undefined,
+        statementPublicationDate: undefined,
+        statementEditionDate: undefined,
+
+        environments: {
+          createMany: {
+            data: originalAudit.environments.map((e) =>
+              omit(e, ["id", "auditUniqueId"])
+            )
+          }
+        },
 
         transverseElementsPage: {
           create: {
@@ -1516,8 +1583,7 @@ export class AuditService {
         ...a.pages.flatMap((p) => p.results)
       ];
 
-      const pagesResults = a.pages.flatMap((p) => p.results);
-      const actualResults = pagesResults.filter((r) =>
+      const actualResults = allResults.filter((r) =>
         CRITERIA_BY_AUDIT_TYPE[a.auditType].find(
           (e) => e.topic === r.topic && e.criterium === r.criterium
         )
@@ -1526,7 +1592,7 @@ export class AuditService {
         actualResults.filter(
           (r) => r.status !== CriterionResultStatus.NOT_TESTED
         ).length /
-        (CRITERIA_BY_AUDIT_TYPE[a.auditType].length * a.pages.length);
+        (CRITERIA_BY_AUDIT_TYPE[a.auditType].length * (a.pages.length + 1));
 
       let complianceLevel = null;
 
@@ -1575,7 +1641,7 @@ export class AuditService {
 
       const auditIsComplete =
         actualResults.length ===
-          CRITERIA_BY_AUDIT_TYPE[a.auditType].length * a.pages.length &&
+          CRITERIA_BY_AUDIT_TYPE[a.auditType].length * (a.pages.length + 1) &&
         actualResults.every((r) => r.status !== "NOT_TESTED");
 
       return {
@@ -1614,7 +1680,7 @@ export class AuditService {
     editUniqueId: string,
     data: UpdateStatementDto
   ): Promise<AuditDto | undefined> {
-    const audit = await this.prisma.audit.update({
+    await this.prisma.audit.update({
       where: { editUniqueId },
       data: {
         initiator: data.initiator,
@@ -1671,6 +1737,8 @@ export class AuditService {
       },
       select: AUDIT_PRISMA_SELECT
     });
+
+    const audit = await this.updateStatementDate(editUniqueId);
 
     return audit;
   }